Rename util in tests to helpers/fixture.ts
CORL-382
