+++ src/core/client/stream/test/helpers/fixture.ts
